+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - fix lost byte in ChunkedDevice (backport of [PR #910])
 - fix director crash on "update slots" when there is a parsing issue with the autochanger or tape devices [PR #919]
 - [Issue #1232]: bareos logrotate errors, reintroduce su directive in logrotate ([backport of PR #918])
+- Fix occasional "NULL volume name" error when non-busy, but blocked drive is unloaded [PR #975]
 
 ### Added
 - packages: Build also for Debian_11 [PR #915]